backend/requirements.txt
arxiv>=2.1.3
python-dotenv>=0.9.9
flask>=3.1.0
flask-cors>=5.0.1
google>=3.0.0
google-generativeai>=0.8.4
googlesearch-python>=1.3.0
phidata>=2.7.10
pycountry>=24.6.1
pypdf>=5.3.0